+++ WIKILocationSelector/Repository/Location.swift
@@ -9,10 +9,10 @@ import Foundation
 
 struct Location: Equatable, Encodable {
     let name: String?
-    let latitude: Double
-    let longitude: Double
+    let latitude: Decimal
+    let longitude: Decimal
     
-    init(name: String? = nil, latitude: Double, longitude: Double) {
+    init(name: String? = nil, latitude: Decimal, longitude: Decimal) {
         self.name = name
         self.latitude = latitude
         self.longitude = longitude
